added delete to cart

+++ app/controllers/cart_controller.rb
@@ -10,9 +10,10 @@ class CartController < ApplicationController
     CartItem.create(cart: current_user.cart, game: game)
   end
 
-  def remove
+  def destroy
     item = CartItem.find(params[:id])
-    item.destroy
+    item.destroy unless item.nil?
+    redirect_to cart_index_path
   end
 
   def create
